Don't overwrite .gitignore if it already exists

+++ gitium/inc/class-git-wrapper.php
@@ -231,7 +231,10 @@ class Git_Wrapper {
 	}
 
 	function init() {
-		file_put_contents( "$this->repo_dir/.gitignore", $this->gitignore );
+		// Don't overwrite .gitignore if it already exists.
+		if ( !file_exists( "$this->repo_dir/.gitignore" ) ) {
+			file_put_contents( "$this->repo_dir/.gitignore", $this->gitignore );
+		}
 		list( $return, ) = $this->_call( 'init' );
 		$this->_call( 'config', 'user.email', 'gitium@presslabs.com' );
 		$this->_call( 'config', 'user.name', 'Gitium' );
